+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 6.7:
+  * Yusuf Ali's English Translation added
+
+## 6.62:
+  * Adds translation labels
+
 ## 6.61:
   * Commentaries integration to global search
   * Global search enhancements, now no need to select layers when performing global search